firmware/python/pyproject.toml
[tool.poetry]
name = "bitkey"
authors = ["Bitkey firmware team <team@bitkey.build>"]
description = "Python client for Bitkey hardware"
version = "0.0.0"
packages = [{ include = "bitkey" }]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
pyserial = "^3.5"
nfcpy = "^1.0.4"
cryptography = "^41.0.2"
sh = "^2.0.4"
pycryptodome = "^3.18.0"
psutil = "^5.9.5"
detools = "^0.51.1"
meson = "^1.2.0"
msgpack = "^1.0.5"
semver = "^3.0.1"
littlefs-python = "^0.6.0"
invoke = "^2.2.0"
bitkey-proto = "*"

# memfault-cli requires an older version of pyelftools
memfault-cli = "1.0.7"
pyelftools = "0.28"

puncover = "0.3.4"
jinja2 = "3.0.0"

# Puncover requires jijna 3.0.0
# Automation
pytest = "^7.4.0"
allure-pytest = "^2.13.2"
bip32 = "^3.4"
bip32utils = "^0.3.post4"
pytest-repeat = "^0.9.1"
pytest-timeout = "^2.1.0"
pygdbmi = "^0.11.0.0"

[tool.poetry.scripts]
bitkey-cli = "bitkey.cli:cli"
bitkey-mfgtest-cli = "bitkey.mfgtest.cli:cli"
bitkey-flash = "bitkey.gdb:cli"

[tool.poetry-dynamic-versioning]
enable = true
style = "semver"
pattern = "^fw-(?P<base>[^+]+)"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
